fix computation backend tests
