+++ pytest.ini
@@ -1,12 +1,12 @@
 [pytest]
-;See link below for available options
-;https://docs.pytest.org/en/latest/reference.html#ini-options-ref
+;See https://docs.pytest.org/en/latest/reference.html#ini-options-ref for available options
 
-markers =
-  large_download
-  slow
-  flaky
-testpaths = tests/
 addopts =
+  # show summary of all tests that did not pass
   -ra
+  # Make tracebacks shorter
   --tb=short
+  # enable all warnings
+  -Wd
+xfail_strict = True
+testpaths = tests
